change handleParameterEvent
